Change relative links under Windows to lower-case
Files under the docset/windows directory are standardized on lower-
case file names (per #42), but have PascalCase in the links. This
means that links do not work correctly on case-sensitive hosts, eg
github. Change the leaves in links to all lower-case to match the
file naming to fix this issue.

docset/mdop and docset/winserver2012r2-ps also require this fix.

docset/winserver2012-ps does NOT need this, as files are named in
PascalCase in that directory, so links work fine as they are.